Réorganisation des screens: nouveau dosier "output" dans client, de "screens" dans "output" et rename output.py en outputFactory.py . (Normalement les imports sont corrigés, mais peut en manquer...)
