+++ client/output/action_bar.py
@@ -2,7 +2,7 @@ from kivy.lang import Builder
 from kivy.properties import StringProperty
 from kivy.uix.boxlayout import BoxLayout
 
-Builder.load_file("action_bar.kv")
+Builder.load_file("client/output/action_bar.kv")
 
 class BoxLayoutWithActionBar(BoxLayout):
     """Outil qui permet de naviguer entre les différents screens,
