+++ client/output/screens/join_game_menu.py
@@ -2,7 +2,7 @@ from kivy.uix.floatlayout import FloatLayout
 from kivy.lang import Builder
 
 
-Builder.load_file("join_game_menu.kv")
+Builder.load_file("client/output/screens/join_game_menu.kv")
 
 class JoinGame(FloatLayout):
     pass
