+++ client/output/screens/layouts.kv
@@ -33,23 +33,6 @@
         size_hint: .2, .06
         on_press: root.main(create=1,join=0)
         
-
-
-
-
-
-
-
-
-
-<UpdateWorlds>:
-    TextInput:
-        pos_hint: {'center_x':.5,'center_y':.5}
-        size_hint: .25, .25
-        text: ""
-        multiline: False
-        on_text_validate: root.on_text_validate(self)
-
     ######################## A décommenter en cas d'extrême désolation dans la recherche de l'easter Egg ########################
     Button:
         pos_hint: {'center_x':.5,'center_y':.1}
